// src/types/OrderSchema.ts
/* tslint:disable */
/* eslint-disable */
/**
 * This file is generated by core-types-ts on behalf of typeconv, DO NOT EDIT.
 * For more information, see:
 *  - {@link https://github.com/grantila/core-types-ts}
 *  - {@link https://github.com/grantila/typeconv}
 */

/** A purchase order for products or services. */
export interface OrderSchema {
    id?: string;
    /**
     * @default
     *     order
     */
    type: string;
    orderNumber?: string;
    status: "draft" | "pendingPayment" | "processing" | "outForDelivery" | "completed" | "paymentFailed" | "cancelled" | "refunded";
    /** User id of purchaser */
    user?: string;
    date?: string | {
        [key: string]: any;
    };
    /**
     * @default
     *     []
     */
    items?: {
        name: string;
        sku?: string;
        /**
         * @default
         *     1
         */
        quantity?: number;
        price: number;
        taxClass?: string;
        deliveryClass?: string;
        [key: string]: any;
    }[];
    /**
     * @default
     *     []
     */
    taxes?: {
        name: string;
        class?: string;
        /**
         * @default
         *     percentage
         */
        rateType?: "percentage" | "flat";
        rate: number;
        total?: number;
        [key: string]: any;
    }[];
    /**
     * @default
     *     []
     */
    fees?: {
        name: string;
        price: number;
        [key: string]: any;
    }[];
    /**
     * @default
     *     []
     */
    delivery?: {
        name: string;
        class?: string;
        /**
         * @default
         *     percentage
         */
        rateType?: "percentage" | "flat";
        rate: number;
        total?: number;
        [key: string]: any;
    }[];
    /**
     * @default
     *     []
     */
    payments?: {
        method?: "folio" | "invoice" | "amex" | "visa" | "masterCard" | "maestro" | "discover" | "diners" | "jcb" | "applePay" | "alipay" | "chinaUnionPay" | "vpay";
        name?: string;
        date?: string | {
            [key: string]: any;
        };
        amount?: number;
        transactionReference?: string;
        [key: string]: any;
    }[];
    [key: string]: any;
}